Fixing TemplatesTest for ruby 1.9.2.
Previously, the test was written to depend on the hash ordering that 1.8.7
produced.  1.9.2 didn't order hash keys the same way, so it was necessary to
break the single assertion up into 4; one assertion for the beginning, one for
the end, and one for each key, value pair.  Now they can be in any order, but
it's still not perfect.