Updating calls to template functions to include options.
Options should be passed to template calls so they can be used, not including
it at first was an oversight.

Options passed to template functions also includes :tree now, which is the
parse tree for the template.  This allows template functions to access the raw
source for the markup, which is needed right now to support syntax
highlighting.