Templates: Added :raw format and format changing.
Now the render options can specify the format for a template's arguments.  This
allows special templates like syntax highlighting to avoid rendering arguments
entirely (by setting options[:template_formats][:template_name] = :raw) so that
the arguments can be rendered by the template itself without side-effects like
link footnotes.  The :raw format simply returns the original parsed text of
arguments.

Another change in this commit is to remove the :tree option that was added a
few commits ago.  It was a bad idea, although necessary in some cases.  I
think, however, for those cases it is better to monkey-patch something in.