Updated template to produce no named args for :raw.
One problem with templates is that if "|" is used in the content, it gets
removed and interpreted as an argument delimiter.  Raw formatted template
arguments are supposed to be a way to increase the flexibility of templates by
making the "raw" parsed text available to templates that wanted to do something
different with it.  For example, syntax highlighting needs to get
the raw content to correctly display code:
 {{ ruby |
 names.each do |name='(unknown)'|
   puts "hello, #{name}!"
 end
 }}

Before this commit, only the value of named parameters was given in raw form,
which meant that in cases like the above, the text could not be reconstructed.
This commit changes the raw format so that all parameters, named and
positional, are given in order as their original text values.