Adding a compromise for template embedding.
Templates can be embedded as inline or block style--just like html.  Inline
style is for options like fixed-width text that can be embedded in a paragraph
without needing a change in the html structure.  Block style is for things that
should not be embedded in paragraphs.  This is now necessary because it is
possible to have markup, not just text, inside a template.

Unfortunately, it is difficult to split paragraphs when a template renders as a
block.  It is also difficult to detect when a template renders as a block,
since anything can be returned by templates when they are rendered.  The
compromise is to assume that templates that contain block values will render as
blocks (since they must) and that templates with only inline values will render
inline.  Templates that contain block arguments are wrapped in divs to avoid
encapsulation problems (paragraphs cannot directly contain paragraphs) and
inline templates are not wrapped.

Text changes:
* Added TextBlock for text blocks that are not paragraphs (in templates)

Language changes:
* Added a paragraph rule that wraps text blocks where they should be
* Changed text rules to use TextBlock
* Similar changes for unsafe_ rules
* Added inline text rules back to templates
* Separated inline and block arguments using classes

Template changes:
* Added Argument subclasses InlineArgument and BlockArgument
* Added block? methods to arguments
* Added block? method to arg lists, true if any arg is a block
* Added block? method to templates, true if the arg list is
* Added block handling to html rendering, which adds a div to block templates