Adding no_wrap? check to paragraphs.
no_wrap? determines whether a paragraph should render a block of text wrapped
in <p> for html.  This inspects the parse tree to see if the block of text is
really just a single template.  If it is a lone template, then it should not be
wrapped since it is most likely a block template.  This is not always the case,
but it should be often enough that it is a good assumption to make.

This ensures that this is not incorrectly wrapped in a paragraph:

{{ floating text box |* list of stuff }}

However, this inline template will be rendered incorrectly, which is part of
the compromise:

{{tt|inline fixed-width text}}

This is a rare situation.