Added argument-safe variants of the markup rules.
The previous attempt to update the language to allow full markup inside
templates (not just a single text paragraph) didn't quite work because it
removed the template_end and arg_delimiter rules from all text. This worked for
templates -- they were correctly matched and could include markup -- but broke
generic markup because out of laziness it included those rules at the top
level. This caused those delimiters to prevent the other markup rules from
correctly matching. For example:

  == Heading }} ==
  Arguments are separated by a "|" character.

The above markup was not matching the heading because headings could not
contain the "}}" delimiter, even outside of templates. Also, the paragraph was
broken into 3 because "|" would be made into its own block. These problems
needed to be resovled by mirroring all of the rules that contain (directly and
transitively) the template end delimiters with template-safe variants. These
all begin with arg_safe_ and are located near the normal rules so they are
edited at the same time.

The new rules use the same classes and modules as the non-safe variants, so
rendering is completely unchanged.